+++ analytics/src/main/java/com/bluetriangle/analytics/Timer.kt
@@ -154,6 +154,10 @@ class Timer : Parcelable {
         null
     )
 
+    private fun isTrackingEnabled():Boolean {
+        return tracker != null
+    }
+
     fun generateNativeAppProperties() {
         nativeAppProperties = NativeAppProperties(
             null,
@@ -231,6 +235,8 @@ class Timer : Parcelable {
      */
     @Synchronized
     fun start(): Timer {
+        if(!isTrackingEnabled()) return this
+
         if (start == 0L) {
             start = System.currentTimeMillis()
             setField(FIELD_UNLOAD_EVENT_START, start)
@@ -262,6 +268,8 @@ class Timer : Parcelable {
      */
     @Synchronized
     fun interactive(): Timer {
+        if(!isTrackingEnabled()) return this
+
         if (start > 0 && interactive == 0L) {
             interactive = System.currentTimeMillis()
             setField(FIELD_DOM_INTERACTIVE, interactive)
@@ -286,6 +294,8 @@ class Timer : Parcelable {
      */
     @Synchronized
     fun end(): Timer {
+        if(!isTrackingEnabled()) return this
+
         if (start > 0 && end == 0L) {
             end = System.currentTimeMillis()
             setField(FIELD_PAGE_TIME, pageTimeCalculator())
@@ -338,6 +348,8 @@ class Timer : Parcelable {
      * Convenience method to submit this timer to the global tracker
      */
     fun submit() {
+        if(!isTrackingEnabled()) return
+
         val tracker = Tracker.instance
         if (tracker != null) {
             if (nativeAppProperties.loadTime == null) {
@@ -711,13 +723,4 @@ class Timer : Parcelable {
         fields[FIELD_ERR] = if (err) "1" else "0"
     }
 
-//    companion object CREATOR : Parcelable.Creator<Timer> {
-//        override fun createFromParcel(parcel: Parcel): Timer {
-//            return Timer(parcel)
-//        }
-//
-//        override fun newArray(size: Int): Array<Timer?> {
-//            return arrayOfNulls(size)
-//        }
-//    }
 }