+++ analytics/src/main/java/com/bluetriangle/analytics/appeventhub/helpers/AppBackgroundNotifier.kt
@@ -1,9 +1,9 @@
-package com.bluetriangle.analytics.launchtime.helpers
+package com.bluetriangle.analytics.appeventhub.helpers
 
 import android.app.Application
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
-import com.bluetriangle.analytics.AppEventHub
+import com.bluetriangle.analytics.appeventhub.AppEventHub
 
 internal class AppBackgroundNotifier(val application: Application):DefaultLifecycleObserver {