+++ analytics/src/main/java/com/bluetriangle/analytics/dynamicconfig/model/BTTRemoteConfigurationMapper.kt
@@ -8,7 +8,6 @@ package com.bluetriangle.analytics.dynamicconfig.model
 import com.bluetriangle.analytics.utility.getBooleanOrNull
 import com.bluetriangle.analytics.utility.getIntOrNull
 import com.bluetriangle.analytics.utility.getJsonArrayOrNull
-import com.bluetriangle.analytics.utility.getStringOrNull
 import org.json.JSONObject
 
 internal object BTTRemoteConfigurationMapper {
@@ -16,6 +15,7 @@ internal object BTTRemoteConfigurationMapper {
     private const val NETWORK_SAMPLE_RATE = "networkSampleRateSDK"
     private const val ENABLE_REMOTE_CONFIG = "enableRemoteConfigAck"
     private const val IGNORE_SCREENS = "ignoreScreens"
+    private const val ENABLE_ALL_TRACKING = "enableAllTracking"
 
     fun fromJson(remoteConfigJson: JSONObject): BTTRemoteConfiguration {
         val networkSampleRate = remoteConfigJson.getIntOrNull(NETWORK_SAMPLE_RATE)?.div(100.0)
@@ -29,10 +29,12 @@ internal object BTTRemoteConfigurationMapper {
                 }
             }
         } ?: listOf()
+        val enableAllTracking = remoteConfigJson.getBooleanOrNull(ENABLE_ALL_TRACKING) ?: true
 
         return BTTRemoteConfiguration(
             networkSampleRate,
             ignoreScreens,
+            enableAllTracking,
             enableRemoteConfig
         )
     }