+++ analytics/src/main/java/com/bluetriangle/analytics/dynamicconfig/model/BTTSavedRemoteConfigurationMapper.kt
@@ -5,6 +5,7 @@
  */
 package com.bluetriangle.analytics.dynamicconfig.model
 
+import com.bluetriangle.analytics.utility.getBooleanOrNull
 import com.bluetriangle.analytics.utility.getDoubleOrNull
 import com.bluetriangle.analytics.utility.getJsonArrayOrNull
 import org.json.JSONArray
@@ -15,6 +16,7 @@ internal object BTTSavedRemoteConfigurationMapper {
     private const val ENABLE_REMOTE_CONFIG = "enableRemoteConfigAck"
     private const val SAVED_DATE = "savedDate"
     private const val IGNORE_SCREENS = "ignoreScreens"
+    private const val ENABLE_ALL_TRACKING = "enableAllTracking"
 
     fun fromJson(jsonObject: JSONObject): BTTSavedRemoteConfiguration {
         val ignoreScreens = jsonObject.getJsonArrayOrNull(IGNORE_SCREENS)?.let { array ->
@@ -31,6 +33,7 @@ internal object BTTSavedRemoteConfigurationMapper {
             jsonObject.getDoubleOrNull(NETWORK_SAMPLE_RATE),
             ignoreScreens,
             jsonObject.getBoolean(ENABLE_REMOTE_CONFIG),
+            jsonObject.getBooleanOrNull(ENABLE_ALL_TRACKING)?: true,
             jsonObject.getLong(SAVED_DATE)
         )
     }
@@ -42,6 +45,7 @@ internal object BTTSavedRemoteConfigurationMapper {
         }
         put(NETWORK_SAMPLE_RATE, config.networkSampleRate)
         put(IGNORE_SCREENS, ignoreListArray)
+        put(ENABLE_ALL_TRACKING, config.enableAllTracking)
         put(ENABLE_REMOTE_CONFIG, config.enableRemoteConfigAck)
         put(SAVED_DATE, config.savedDate)
     }